+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - add option to approximate elevation of new point
   [#18](https://gitlab.com/eneiluj/gpxedit-oc/issues/18) @TheTiPi
+- support for ownCloud admin settings
 
 ### Changed
 - change way to cut lines, hover the middle marker and press Del